LayoutTests/imported/w3c/web-platform-tests/css/css-align/blocks/justify-self-text-align-2-expected.txt

normal
auto
left

PASS .item 1
FAIL .item 2 assert_equals:
<div class="container">
  <div class="item" style="justify-self: normal;" data-offset-x="100">
    normal
  </div>
  <div class="item" style="justify-self: auto;" data-offset-x="50">
    auto
  </div>
  <div class="item" style="justify-self: left;" data-offset-x="0">
    left
  </div>
</div>
offsetLeft expected 50 but got 100
FAIL .item 3 assert_equals:
<div class="container">
  <div class="item" style="justify-self: normal;" data-offset-x="100">
    normal
  </div>
  <div class="item" style="justify-self: auto;" data-offset-x="50">
    auto
  </div>
  <div class="item" style="justify-self: left;" data-offset-x="0">
    left
  </div>
</div>
offsetLeft expected 0 but got 100
